Fix URLs, modernize code (Codable, Swift 5), better error reporting
This updates the hardcoded URLs to make the script work again.

The main change is to switch to Codable for decoding the JSON downloaded from data.swift.org and for encoding the JSON response to Alfred.

To decouple the actual JSON formats used by the source (swift.org) and the destination (Alfred), this introduces two new structs, ProposalDTO and AlfredItem. These model the data formats we need to work with faithfully. The Proposal struct still exists as the currency type for our own business logic. So the program now works like this:

Download and parse [ProposalDTO] → Convert to [Proposal] and process → Convert to [AlfredItem] and return

The new structs mirror the data formats 1:1, which makes them a good piece of documentation. It should also be easy to support more fields, both in the input and in the output format.