Don’t repeat proposal title in Alfred’s subtitle field
We had previously repeated the proposal title in the subtitle because the title field might get truncated earlier due to its bigger font size. However, now that we display some additional metadata in the subtitle (version when a proposal was implemented, review period of proposals in review), I think not repeating the title is better.